requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyio==4.0.0
APScheduler==3.10.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
emoji-country-flag==1.3.2
exceptiongroup==1.1.3
frozenlist==1.4.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
multidict==6.0.5
mysql-connector-python==9.0.0
persiantools==3.0.1
python-coinmarketcap==0.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-telegram-bot==20.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
requests==2.31.0
rfc3986==1.5.0
setuptools==70.0.0
six==1.16.0
sniffio==1.3.0
tornado==6.4.1
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
yarl==1.9.4
